Initial commit of lambda function 'dashboardQueries'.
This function provides an interface between the dashboard web app and
the Postgresql database. The function accepts a simplified query, parses
it, validates it against pre-defined allowed queries, and then executes
the real Postgresql query. Results are returned in a .csv format.

Because lambda does not nativly support querying Postgresql, we need to
include the psycopg2 package in the upload. And because that package
includes an architecture specific binary, we must tweak the packages
before we zip and upload them.

For convenience, a makefile is also included to create the appropriate
.zip file, and to upload it to AWS.