Initial tblHelper for automatic TB SRN management.
Initial checkin of lambda code for automatic TBS management.

TB-Loaders can call the "reserve" function to obtain a block of TB serial numbers.
The signed-in user's email is used to identify the user, and if they are new to
TB serial numbers, a new TB loader id is allocated for them.

Subsequently, they can allocate blocks of serial numbers, until the range (~64k)
is exhausted. At that point, a new TB loader id will be allocated to the user.

Given that we have ~3000 unassigned Talking Books, I do not expect this to
ever be needed, even if a single TB-Loader user initializes ALL of the Talking
Books..

If the user signs into a new device or laptop, a new block will be allocated. If
for some reason the old block is lost (reformatted hard drive, new phone, ...),
a new block is allocated. This is wasteful of serial numbers, but avoids the
possibility of duplicates.