Initial checkin of sfToAws.
When a program is created or updated in Salesforce, a .json file is
written to S3 and this application is triggered. It will read the .json
file and create the entries in RDS, S3, and Dyamodb for the new program.

At least that's the plan. For the time being, it creates and emails a
command line to invoke newAcm to create the new ACM database. Once I
trust the data and flow better, this can become fully automatic.