Initial checkin of twnxHelper.
This is a utility to serve an image (a preview) or a .twbx file for a
program. The data is simply served as a signed URL to the object in S3.

The utility can also upload a file, but that's redundant and should be
removed; use the aws cli instead.