Add deployment_uuid and deployment_timestamp to usage queries.
When a deployment is placed onto a Talking Book, a deployment.properties file
is created on the Talking Book. Two of the fields in that file are the
time that the deployment is performed and a UUID of that deployment. This
update makes those available to usage queries. (The UUID lets us work around
duplicate Talking Book IDs.)